New Code Developed
Currently Implemented and tested...
Solenoid Timer Control
Pressure Sensor
Temperature Sensor
UV Data

+++ rocksat/pressure.h
@@ -1,7 +1,7 @@
 #ifndef PRESSURE_H_   /* Include guard */
 #define PRESSURE_H_
 
-void getPressureTempData(short*);
+void getPressureTempData(float*);
 
 
 #endif
